New function completed

+++ ComplexPlane.h
@@ -1,5 +1,7 @@
 #pragma once
 #include <SFML/Graphics.hpp>
+#include <SFML/System/Thread.hpp>
+#include <cmath>
 #include <iostream>
 using namespace sf;
 using namespace std;
@@ -36,5 +38,5 @@ class ComplexPlane : public Drawable
 
 		int countIterations(Vector2f coord);
 		void iterationsToRGB(size_t count, Uint8& r, Uint8& g, Uint8& b);
-		Vector2f mapPixelToCoords(Vector2f mousePixel);
+		Vector2f mapPixelToCoords(Vector2i mousePixel);
 };
